backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
phonenumbers==8.13.19
azure-identity==1.14.0
azure-mgmt-resource==23.0.1
msal==1.24.0
